seperate and main macro
